Update .gitignore
The project presents only the client part with the simulation of receiving data from the server through a fetch request (MyTeg component) and when the order is closed, the final information in the specified form is printed from the console. The application implements functionality - adding a client, adding an order (for one client, you can place several orders with combining tables), automatic seating of clients when loading a file with clients for the first time, closing an order after 90 seconds, when receiving an order, the color of the table changes to red and after 60 seconds to orange, when you click on an occupied table, a modal window pops up with information about the client.